2.4-celery/upscale/requirements.txt
opencv-contrib-python
opencv-python
redis
pymongo
psycopg2-binary
cachetools
celery
nanoid
SQLAlchemy
singleton-decorator
python-dotenv
Flask
Flask-PyMongo
gunicorn
requests
